Modificacion de las rutas de imagenes

+++ app/Http/Controllers/MaterialController.php
@@ -42,7 +42,7 @@ class MaterialController extends Controller
         $request->validate([
             'name' => ['required', 'string', 'max:255', 'regex:/^[\p{L}\s]+$/u'],
             'image' => ['required', 'image', 'mimes:jpeg,png,jpg,gif', 'max:5120'],
-            'description' => ['required','string','max:255','regex:/^[\p{L}\s]+$/u',],
+            'description' => ['required', 'string', 'max:255', 'regex:/^[\p{L}\s]+$/u',],
             'material_category_id' => ['required', 'exists:material_categories,id'],
             'weight' => ['required', 'numeric', 'regex:/^\d+(\.\d{1,2})?$/'],
             'value' => ['required', 'numeric', 'regex:/^\d+(\.\d{1,2})?$/'],
@@ -53,12 +53,13 @@ class MaterialController extends Controller
 
         if ($request->hasFile('image')) {
             $image = $request->file('image');
-            $imageName = time() . '_' . $image->getClientOriginalName(); 
-            $image->move(public_path('material_image'), $imageName);
+            $imageName = time() . '_' . $image->getClientOriginalName();
+            $image->move(base_path('../material_image'), $imageName);
             $data['image'] = '/material_image/' . $imageName;
         }
 
         Material::create($data);
+
 
         return redirect()->route('materials.index')->with('success', 'Material creado exitosamente.');
     }
@@ -76,7 +77,7 @@ class MaterialController extends Controller
         $request->validate([
             'name' => ['required', 'string', 'max:255', 'regex:/^[\p{L}\s]+$/u'],
             'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif', 'max:5120'],
-            'description' => ['required','string','max:255','regex:/^[\p{L}\s]+$/u',],
+            'description' => ['required', 'string', 'max:255', 'regex:/^[\p{L}\s]+$/u',],
             'material_category_id' => ['required', 'exists:material_categories,id'],
             'weight' => ['required', 'numeric', 'regex:/^\d+(\.\d{1,2})?$/'],
             'value' => ['required', 'numeric', 'regex:/^\d+(\.\d{1,2})?$/'],
@@ -86,17 +87,16 @@ class MaterialController extends Controller
 
         if ($request->hasFile('image')) {
             if ($material->image) {
-                $oldImagePath = public_path('material_image/' . basename($material->image));
+                $oldImagePath = base_path('../material_image/' . basename($material->image));
                 if (file_exists($oldImagePath)) {
                     unlink($oldImagePath);
                 }
             }
-    
-           
+
             $image = $request->file('image');
-            $imageName = time() . '_' . $image->getClientOriginalName(); 
-            $image->move(public_path('material_image'), $imageName); 
-    
+            $imageName = time() . '_' . $image->getClientOriginalName();
+            $image->move(base_path('../material_image'), $imageName);
+
             $data['image'] = '/material_image/' . $imageName;
         } else {
             $data['image'] = $material->image;
@@ -104,20 +104,28 @@ class MaterialController extends Controller
 
         $material->update($data);
 
+
         return redirect()->route('materials.index')->with('success', 'Material actualizado exitosamente.');
     }
 
 
     public function destroy(Material $material)
     {
+
+        if ($material->elements()->count() > 0) {
+            return redirect()->route('materials.index')
+                ->with('error', 'No se puede eliminar el material porque está asociado a elementos.');
+        }
+        
         if ($material->image) {
-            $imagePath = public_path('material_image/' . basename($material->image));
+            $imagePath = base_path('../material_image/' . basename($material->image));
             if (file_exists($imagePath)) {
                 unlink($imagePath);
             }
         }
-        
+
         $material->delete();
+
         return redirect()->route('materials.index')->with('success', 'Material eliminado exitosamente.');
     }
 }
